pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "litellm>=1.70.2",
    "mcp>=1.9.0",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
]
